(PC-22512) refactor(offer): only venue modal title and toggle sticky
